feat(arch): refactor using gradle init
