+++ .gitignore
@@ -40,3 +40,6 @@ out/
 
 ### VS Code ###
 .vscode/
+
+# Ignore Gradle build output directory
+build
