+++ app/src/main/java/com/pelukron/App.java
@@ -1,0 +1,21 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package com.pelukron;
+
+public class App {
+    public String getGreeting() {
+        return "Hello World!";
+    }
+
+    public static void main(String[] args) {
+        System.out.println(new App().getGreeting());
+
+        try {
+            Lista list = new Lista();
+            list.menu();
+        } catch (Exception e) {
+            System.err.println(e);
+        }
+    }
+}
