+++ app/src/main/java/com/pelukron/Nodo.java
@@ -2,7 +2,7 @@
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.pelukron.demo;
+package com.pelukron;
 
 /**
  *
