--- src/test/java/com/pelukron/demo/DemoApplicationTests.java
@@ -1,9 +1,0 @@
-package com.pelukron.demo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DemoApplicationTests {
-
-}
